C#’s program control statements.
In The Name Of GOD
C#’s program control statements.
There are three categories of program control statement:
1.Selection statement : if & switch.
2.Iteration statement : for , while , do-while & foreach.
3.Jump statement : break , continue , goto , return & throw.

if(condition)
 statement; 
else if(condition)
 statement;
 else if(condition) 
statement;
 . . .
 else 
statement;
 the switch statement :
switch(expression)
 {  
case constant1: 
      statement sequence  
     break;
   case constant2:   
 statement sequence  
      break; 
  case constant3:   
 statement sequence      
  break;
  .  .  .  
default: 
statement sequence
 break;
 }
The switch expression must be of an integer type, such as char, byte, short, or int, of an enumeration type, or of type string. (Enumerations and the string type are described later in this book.) Thus, floating-point expressions, for example, are not allowed. Frequently, the expression controlling the switch is simply a variable. The case constants must be of a type compatible with the expression. No two case constants in the same switch can have identical values. The default sequence is executed if no case constant matches the expression. The default is optional; if it is not present, no action takes place if all matches fail. When a match is found, the statements associated with that case are executed until the break is encountered.

for(initialization; condition; iteration) 
{
 statement sequence
 }
……..
for(;;)
 // intentionally infinite loop
 { 
 //...
 }
while(condition)
 statement;
do
 {   
statements;
 }
 while(condition);
The foreach loop cycles through the elements of a collection. A collection is a group of objects. C# defines several types of collections, of which one is an array.
Using break to Exit a Loop